+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ## Unreleased
   - Allow arbitrary map sizes below 256x256
   - Draw a higher details minimap for map sizes below 32x32
-  - Implement a `</PLAYSOUND xxx>` tag for credits/establishing shots
+  - Implement new tags for credits/establishing shots:
+    * `</PLAYSOUND xxx>`
+    * `</SCREENUPPERLEFT>`
+    * `</SCREENUPPERRIGHT>`
+    * `</SCREENLOWERRIGHT>`
   - Fix wrong colors in magma/water tiles on Ashworld/Ice tilesets (by DarkenedFantasies)
 
 ## 0.1.0 - Anxious Arbiter